Refactor WriterGroupPublishState out of IPubSubConnection file
